+++ be1-go/message2/messagedata/meeting_state.go
@@ -8,11 +8,11 @@ type MeetingState struct {
 	Action                 string                  `json:"action"`
 	ID                     string                  `json:"id"`
 	Name                   string                  `json:"name"`
-	Creation               int                     `json:"creation"`
-	LastModified           int                     `json:"last_modified"`
+	Creation               int64                   `json:"creation"`
+	LastModified           int64                   `json:"last_modified"`
 	Location               string                  `json:"location"`
-	Start                  int                     `json:"start"`
-	End                    int                     `json:"end"`
+	Start                  int64                   `json:"start"`
+	End                    int64                   `json:"end"`
 	Extra                  json.RawMessage         `json:"extra"`
 	ModificationID         string                  `json:"modification_id"`
 	ModificationSignatures []ModificationSignature `json:"modification_signatures"`
