+++ be1-go/message2/test/messagedata/roll_call_close_test.go
@@ -31,7 +31,7 @@ func Test_Roll_Call_Close(t *testing.T) {
 	require.Equal(t, "close", msg.Action)
 	require.Equal(t, "XXX", msg.UpdateID)
 	require.Equal(t, "XXX", msg.Closes)
-	require.Equal(t, 123, msg.ClosedAt)
+	require.Equal(t, int64(123), msg.ClosedAt)
 	require.Len(t, msg.Attendees, 1)
 	require.Equal(t, "XXX", msg.Attendees[0])
 }
